alternate source adding change. change masktype to uint32_t.
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@529 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ Flags.h
@@ -5,7 +5,7 @@ namespace dcpp {
 
 class Flags {
 	public:
-		typedef uint16_t MaskType;
+		typedef uint32_t MaskType;
 
 		Flags() : flags(0) { }
 		Flags(const Flags& rhs) : flags(rhs.flags) { }
